refactor: Outsource large biz logic from vuex store
